Mqtt5AsyncClientModule refactor
Changed the default values that are used for selected broker and port
when they are absent from shared preferences. Removed port range
checking since the values should be validated before being accepted as
shared preference inputs. Added a shared preferences value that can be
keyed: broker_state. This will be leveraged to convey the current state
of the client.